Inline ckbtc_minter_kyt
